requirements.txt
# Install these dependencies
transformers
torch==2.0
chromadb
sentence-transformers 
huggingface_hub 
chainlit 
mistralai 
python-dotenv 
networkx 
matplotlib 



pydantic==2.10.1 